[chapter04] chapter04 파일 수정
Former-commit-id: 76b3f723373bf8fc1aaf29277a940a79b7f7954a

+++ chapter04/src/user/service/UserDeleteService.java
@@ -12,8 +12,14 @@ public class UserDeleteService implements UserService {
 		System.out.println("삭제할 계정의 id를 입력하세요");
 		Scanner sc = new Scanner(System.in);
 		String id = sc.nextLine();
-		userDAO.deleteUser(id);
-		System.out.println("유저의 정보가 삭제되었습니다.");
+		int su = userDAO.getUser(id);
+		if(su==1) {
+			userDAO.deleteUser(id);
+			System.out.println("유저의 정보가 삭제되었습니다.");
+		} else {
+			System.out.println("유저의 정보가 존재하지 않습니다.");
+		}
+		
 	}
 
 }
